AppRTC speed improvements, default to HD
Turn on HD by default on Chrome desktop.
Turn on improved BWE by default.
Allow the start bitrate to be set from a URL param.
Allow the ICE transports value to be set from a URL param.
Speed up the call connect transition.
Rewrite a bunch of SDP-munging JS to be simpler.