Support for filing bugs.
When being forwarded to the issue tracker, the contents of the issue
to be filed is not lost if the user has to authenticate first.
Details on fields that can be pre-populated:
https://code.google.com/p/support/wiki/IssueTracker#Issue_entry
It seems hard to auto-attach a file.

A future improvement can be to intercept all console logging and
store a copy of each line in a texttarea element. Then it could
be included in the description of the bug to be filed.

TESTED=Ran the browser detection part with Chrome, Firefox, IE and Safari on desktop to ensure it was working properly.