Analytics class API for impressions to BigQuery
The Analytics class provides an API for sending impressions to
BigQuery. The bulk of the logic provided is to handle cases where apprtc
is run on AppEngine vs when it is running locally for dev.

When running on AppEngine, Analytics uses AppAssertionCredentials which
automatically provides valid credentials for the AppEngine project.

If running locally nothing is logged to BigQuery by default. In the case
that a 'secrets.json' file is provided all logging will be sent to
that project specified in the secrets file.