e2ee: polyfill RTCEncoded(Audio|Video)Frame.getMetadata()
For the encoded video frame this is an incomplete polyfill since
the additional properties would have to be parsed from the additionalData
and depends on the codec used.

In workers, polyfilling is not possible since self.RTCEncodedAudioFrame does
not exist. The polyfill is done at runtime as another transformation step.