Make the gDM example request audio
As a Chromium committer, I use this page often to trigger gDM. I know many others who do. And we often need audio, which then requires us to use the DevTools console. Or worse, we interact with someone trying to reproduce an issue, and they report that "audio also stopped working" after interacting with this page.