add interop tests
adds interoperability tests between Chrome unstable and Firefox
Nightly. These tests are designed to run nightly as github actions.

They do not rely on a signaling server, instead use the mocha-based
test that is controlling the individual selenium webdriver instances
to act as a signaling channel which exchanges offers and answers with
candidates.
This pattern is described in the testbed repository
  https://github.com/fippo/testbed
from 2016, this is a "more modern" take on the same subject.

To run the tests locally,
* npm install --no-save chromedriver geckodriver
* npm run mocha test/interop/connection